rename, augment coreDatum testing
